fixed restart bug
