+++ challenge.js
@@ -1,32 +1,8 @@
 var previousId, currentId, currentRow, currentCol, numLeft = 20, fadeOutIndex = 0;
-var didStart = false, didEnd = false;
+var didStart = false, didEnd = false, canRestart = true;
 
 $('.row-lost').children().hide();
 $('.row-won').children().hide();
-
-//jQuery.fn.reverse = [].reverse;
-
-jQuery.fn.reverse = function() {
-  return this.pushStack(this.get().reverse(), arguments);
-};
-
-jQuery.fn.removeClassExcept = function (val) {
-    return this.each(function (index, el) {
-        var keep = val.split(" "),  // list we'd like to keep
-        reAdd = [],          // ones that should be re-added if found
-        $el = $(el);       // element we're working on
-        // look for which we re-add (based on them already existing)
-        for (var c = 0; c < keep.length; c++){
-          if ($el.hasClass(keep[c])) reAdd.push(keep[c]);
-        }
-
-        // drop all, and only add those confirmed as existing
-        $el
-          .removeClass()               // remove existing classes
-          .addClass(reAdd.join(' '));  // re-add the confirmed ones
-    });
-};
-
 
 //.addClass("current");
 
@@ -105,20 +81,24 @@ var addToChosen = function() {
 }
 
 var youLose = function() {
+  canRestart = false;
   $('.logo-background').fadeOut();
   fadeOutElements('.row3');
   fadeOutElements('.row2');
   fadeOutElements('.row1');
   fadeInElements('.row-lost');
+  window.setTimeout(allowRestart, 3600);
 }
 
 var youWin = function() {
+  canRestart = false;
   $('.logo-lose').fadeOut();
   $('.logo-background').fadeOut();
   fadeOutElements('.row3');
   fadeOutElements('.row2');
   fadeOutElements('.row1');
   fadeInElements('.row-won');
+  window.setTimeout(allowRestart, 3600);
 }
 
 var fadeOutElements = function(specificClass) {
@@ -136,22 +116,24 @@ var fadeInElements = function(specificClass) {
   });
 }
 
-
-var showInstructions = function(){
-  alert();
+var allowRestart = function(){
+  canRestart = true;
 }
 
-
 var restart = function(){
-  numLeft = 20, fadeOutIndex = 0, didStart = false, didEnd = false;
-  $('.row-lost').children().hide();
-  $('.row-won').children().hide();
-  $('.row1').children().show();
-  $('.row2').children().show();
-  $('.row3').children().show();
-  $('.logo-lose').fadeOut();
-  $('.logo-background').fadeIn();
-  $('.square').removeClassExcept('square');
+
+  if (canRestart == true) {
+    numLeft = 20, fadeOutIndex = 0, didStart = false, didEnd = false;
+    $('.row-lost').children().hide();
+    $('.row-won').children().hide();
+    $('.row1').children().show();
+    $('.row2').children().show();
+    $('.row3').children().show();
+    $('.logo-lose').fadeIn();
+    $('.logo-background').fadeIn();
+    $('.square').removeClassExcept('square');
+    $('.row-won').children().addClass('selected');
+  }
 }
 
 $(".restart").click(function(){
