+++ index.css
@@ -94,6 +94,7 @@ body {
   top: 20px;
   padding: 20px 20px 20px 20px;
   border-radius: 3px
+  width: 200px;
 }
 
 #headline-container, #links-container{
